[ALPS04239425] platform: add k63v1_bsp project
Create k63v1_bsp project

MTK-Commit-Id: c1d0030f8e999745df759b797e1b9da07ac61f57

CR-Id: ALPS04239425
Feature: Configuration

+++ drivers/misc/mediatek/video/mt6763/dispsys/display_recorder.c
@@ -1039,7 +1039,7 @@ void caculate_fps(void)
 	struct dprec_logger_fps *l = &logger_fps;
 	int i;
 	unsigned long long fps_high_tmp = 0;
-	unsigned long fps_low_tmp = 0;
+	unsigned long long fps_low_tmp = 0;
 	unsigned long long total = 0;
 
 	total = l->ts_end_update - l->ts_start_update;
@@ -1078,7 +1078,7 @@ int dprec_logger_get_result_value(enum DPREC_LOGGER_ENUM source,
 	int len = 0;
 	struct dprec_logger *l = &logger[source];
 	unsigned long long fps_high = 0;
-	unsigned long fps_low = 0;
+	unsigned long long fps_low = 0;
 	unsigned long long avg;
 	unsigned long long count;
 	unsigned long long total = 0;
